Move UserProperty to more appropriately named file
